+++ lib/lists.gd
@@ -3,7 +3,7 @@
 #W  lists.gd                    GAP4 package `Utils'               Stefan Kohl
 ##                                                               Chris Wensley
 ##
-#Y  Copyright (C) 2015-2017, The GAP Group 
+#Y  Copyright (C) 2015-2018, The GAP Group 
 
 if OKtoReadFromUtils( "XMod" ) then
 
@@ -15,6 +15,25 @@ if OKtoReadFromUtils( "XMod" ) then
 DeclareOperation( "PrintListOneItemPerLine", [ IsList ] );
 
 fi; 
+
+##############################################################################
+##  added global function and method for iterators to complement the above 
+##  
+#F  PrintOneItemPerLine( <args> ) 
+#O  PrintIteratorOneItemPerLine( <iter> ) 
+##  
+DeclareGlobalFunction( "PrintOneItemPerLine" );
+DeclareOperation( "PrintIteratorOneItemPerLine", [ IsIterator ] );
+
+##############################################################################
+##  added global function and methods to print subsets of lists/iterators  
+##  
+DeclareGlobalFunction( "PrintSelection" ); 
+DeclareOperation( "PrintSelectionFromList", [ IsList, IsPosInt, IsPosInt ] ); 
+DeclareOperation( "PrintSelectionFromIterator", 
+    [ IsIterator, IsPosInt, IsPosInt ] ); 
+DeclareOperation( "PrintSelectionFromListByList", [ IsList, IsList ] ); 
+DeclareOperation( "PrintSelectionFromIteratorByList", [ IsIterator, IsList ] ); 
 
 if OKtoReadFromUtils( "ResClasses" ) then