+++ CHANGES.md
@@ -1,7 +1,9 @@
 # CHANGES log for the 'Utils' package
 
-## Version 0.51 for GAP 4.8.9 (08/01/18) 
+## Version 0.51 for GAP 4.8.10 (21/01/18) 
 
+ * (21/01/18) PrintOneItemPerLine plus alternative method for iterators 
+              added function PrintSelection(L,first,step) for lists/iterators
  * (06/01/18) rebuilt the manual using the AutoDoc package 
  * (22/12/17) removed examples/ folder